Sync

+++ Data Structures/HashTable.c
@@ -23,7 +23,7 @@ typedef struct entry
 
 int hash(char*key)
 {
-	int i, len = strlen(key), h = 0;
+	int i, len = (int)strlen(key), h = 0;
 	for (i = 0; i < len; i++)
 	{
 		h += pow(key[i], i);
